+++ src/graphql_clj/validator/rules/variables_in_allowed_position.clj
@@ -34,6 +34,8 @@
    If both are list types, the variable item type can be more strict
    than the expected item type (contravariant)."
   [s maybe-sub-type super-type]
+  (assert maybe-sub-type)
+  (assert super-type)
   (cond (= maybe-sub-type super-type)                                                          true   ;; Equivalent type is a valid subtype
         (and (required-type? super-type) (not (required-type? maybe-sub-type)))                false  ;; If superType is non-null, maybeSubType must also be non-null.
         (not= (list-type? maybe-sub-type) (list-type? super-type))                             false  ;; If superType is not a list, maybeSubType must also be not a list, and vice versa.
@@ -49,11 +51,12 @@
 
 (defnodevisitor variable-type-mismatch :post :argument
   [{:keys [spec variable-name] :as n} s]
-  (let [arg-type (s/get-spec spec)
-        var-spec (spec/spec-for {:node-type :variable-usage :variable-name variable-name} s)
-        var-type (s/get-spec var-spec)
-        var-def  (spec/get-type-node var-spec s)]
-    (when-not (subtype-of s (effective-type var-type var-def) arg-type)
-      {:state (ve/update-errors s (variable-type-error var-def n s))})))
+  (when variable-name
+    (let [arg-type (s/get-spec spec)
+          var-spec (spec/spec-for {:node-type :variable-usage :variable-name variable-name} s)
+          var-type (s/get-spec var-spec)
+          var-def (spec/get-type-node var-spec s)]
+      (when-not (subtype-of s (effective-type var-type var-def) arg-type)
+        {:state (ve/update-errors s (variable-type-error var-def n s))}))))
 
 (def rules [variable-type-mismatch])